# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Smriti API
  description: Smarter Home for all your Photos and Videos
  version: 24.06.24
servers:
  - url: https://localhost:5001
...
paths:
...
  /v1/mediaItems:
    get:
      tags:
        - MediaItems
      parameters:
        - name: page
          in: query
          description: Pagination page
          schema:
            type: number
            default: 1
        - name: limit
          in: query
          description: Pagination limit
          schema:
            type: number
            default: 15
        - name: type
          in: query
          description: MediaItem Type
          schema:
            type: string
            example: video
        - name: category
          in: query
          description: MediaItem Category
          schema:
            type: string
            example: live
        - name: status
          in: query
          description: MediaItem Status
          schema:
            type: string
            example: FAILED
      summary: Get all existing mediaItems
      description: Get all existing mediaItems
      operationId: getMediaItems
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaItems'
        '400':
          description: Bad Request
        '401':
          description: Unauthenticated
        '500':
          description: Internal Server Error
    post:
      tags:
        - MediaItems
      summary: Upload a new mediaItem
      description: Upload a new mediaItem
      operationId: uploadMediaItem
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Created
        '400':
          description: Bad Request
        '401':
          description: Unauthenticated
        '500':
          description: Internal Server Error
...
  /v1/favourites:
    get:
      tags:
        - Library
      summary: Get all favourite mediaItems
      description: Get all favourite mediaItems
      operationId: getFavouriteMediaItems
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaItems'
        '400':
          description: Bad Request
        '401':
          description: Unauthenticated
        '500':
          description: Internal Server Error
    post:
      tags:
        - Library
      summary: Add favourite mediaItems
      description: Add favourite mediaItems
      operationId: addFavouriteMediaItems
      security:
        - bearerAuth: []
      requestBody:
        description: List of mediaitems request
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaItemsRequest'
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Unauthenticated
        '500':
          description: Internal Server Error
    delete:
      tags:
        - Library
      summary: Remove favourite mediaItems
      description: Remove favourite mediaItems
      operationId: removeFavouriteMediaItems
      security:
        - bearerAuth: []
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Unauthenticated
        '500':
          description: Internal Server Error
  /v1/hidden:
    get:
      tags:
        - Library
      summary: Get all hidden mediaItems
      description: Get all hidden mediaItems
      operationId: getHiddenMediaItems
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaItems'
        '400':
          description: Bad Request
        '401':
          description: Unauthenticated
        '500':
          description: Internal Server Error
    post:
      tags:
        - Library
      summary: Add hidden mediaItems
      description: Add hidden mediaItems
      operationId: addHiddenMediaItems
      security:
        - bearerAuth: []
      requestBody:
        description: List of mediaitems request
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaItemsRequest'
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Unauthenticated
        '500':
          description: Internal Server Error
    delete:
      tags:
        - Library
      summary: Remove hidden mediaItems
      description: Remove hidden mediaItems
      operationId: removeHiddenMediaItems
      security:
        - bearerAuth: []
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Unauthenticated
        '500':
          description: Internal Server Error
  /v1/trash:
    get:
      tags:
        - Library
      summary: Get all deleted mediaItems
      description: Get all deleted mediaItems
      operationId: getDeletedMediaItems
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaItems'
        '400':
          description: Bad Request
        '401':
          description: Unauthenticated
        '500':
          description: Internal Server Error
    post:
      tags:
        - Library
      summary: Add deleted mediaItems
      description: Add deleted mediaItems
      operationId: addDeletedMediaItems
      security:
        - bearerAuth: []
      requestBody:
        description: List of mediaitems request
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaItemsRequest'
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Unauthenticated
        '500':
          description: Internal Server Error
    delete:
      tags:
        - Library
      summary: Remove deleted mediaItems
      description: Remove deleted mediaItems
      operationId: removeDeletedMediaItems
      security:
        - bearerAuth: []
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Unauthenticated
        '500':
          description: Internal Server Error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic
  schemas:
...
    MediaItem:
      type: object
      properties:
        id:
          readOnly: true
          type: string
          format: uuid
          example: 102dc522de-1669-48a4-8270-4950a0237599
        userId:
          readOnly: true
          type: string
          format: uuid
          example: 102dc522de-1669-48a4-8270-4950a0237599
        filename:
          readOnly: true
          type: string
          example: IMG_4942.jpg
        description:
          readOnly: true
          type: string
          example: Picture of a pizza from New York
        mimeType:
          readOnly: true
          type: string
          example: image/jpeg
        placeholder:
          readOnly: true
          type: string
          example: base-64-placeholder-image
        sourceUrl:
          readOnly: true
          type: string
          example: https://cdn.smriti/source
        previewUrl:
          readOnly: true
          type: string
          example: https://cdn.smriti/preview
        thumbnailUrl:
          readOnly: true
          type: string
          example: https://cdn.smriti/thumbnail
        favourite:
          type: boolean
          example: false
        shared:
          type: boolean
          example: true
        hidden:
          type: boolean
          example: false
        deleted:
          type: boolean
          example: false
        status:
          readOnly: true
          type: string
          enum: ["UNSPECIFIED", "PROCESSING", "READY", "FAILED"]
        mediaItemType:
          readOnly: true
          type: string
          enum: ["photo", "video"]
        mediaItemCategory:
          readOnly: true
          type: string
          enum: ["default", "screenshot", "panorama", "slow", "motion", "live"]
        width:
          readOnly: true
          type: integer
          example: 1080
        height:
          readOnly: true
          type: integer
          example: 720
        creationTime:
          readOnly: true
          type: string
          format: date-time
          example: "2022-09-14T16:58:28Z"
        latitude:
          readOnly: true
          type: number
          example: 19.384982
        longitude:
          readOnly: true
          type: number
          example: -70.638324
        cameraMake:
          readOnly: true
          type: string
          example: Apple
        cameraModel:
          readOnly: true
          type: string
          example: iPhone 12
        focalLength:
          readOnly: true
          type: string
          example: 5.49 mm
        apertureFNumber:
          readOnly: true
          type: string
          example: f/1.8
        isoEquivalent:
          readOnly: true
          type: string
          example: ISO758
        exposureTime:
          readOnly: true
          type: string
          example: 1/20
        fps:
          readOnly: true
          type: string
          example: 60
        createdAt:
          readOnly: true
          type: string
          format: date-time
          example: "2022-09-14T16:58:28Z"
        updatedAt:
          readOnly: true
          type: string
          format: date-time
          example: "2022-09-14T16:58:28Z"
...
    MediaItemsRequest:
      type: object
      properties:
        mediaItems:
          type: array
          items:
            type: string
            format: uuid
            example: 102dc522de-1669-48a4-8270-4950a0237599
...
    MediaItems:
      type: array
      items:
        $ref: '#/components/schemas/MediaItem'
...